Can't find Mono directory Fix on OSX
The latest Unity versions seem to have a different folder structure. I found a Mono bin folder under "MonoBleedingEdge/bin" and added that to the code, but I'm not 100% sure this really solves the problem.

I also changed the way the environment variables are set. Before, the environment variables were added each time the AssetProcessor ran, even if the entry already existed. Now we check if there's already an entry before adding a duplicate.